feat: Efficient snapshots
This needs a bit of explaining as it deviates from the behavior of the official (closed source) implementation.

As notes are end to end encyrpted, it is impossible to do incremental backups without a change in client architecture (sending only the diff, restoring based on each diff like git)

This approach takes up a lot of space, especially when files get large. This means each small change (e.g. 1 character) will require backing up the *whole* file.

Instead, the approach here is to give you access to this on a per instance basis. I assume you will probably need to inspect and recover minor changes during a specific session and will only close the editor once things are done.

In  this case, a snapshot is made every time the app is closed, deleting the small changes but keeping a copy of the final product of each instance.